fix: デプロイエラーを修正
- テストのセットアップファイルパスを修正
- generateWithReplicateの引数エラーを修正
- vitest.config.tsのsetupFilesパスを正しいものに変更
- ESLintエラーを修正（未使用のインポートを削除、any型を型定義に置き換え）

テストが全て通るようになりました。
