+++ frontend/src/components/__tests__/ImageGenerationSection.test.tsx
