+++ frontend/src/components/steps/__tests__/CategoryStep.category-change.test.tsx
