+++ frontend/src/components/steps/__tests__/CategoryStep.test.tsx
