+++ frontend/src/components/steps/__tests__/StyleStep.test.tsx
