+++ frontend/src/services/__tests__/commercialImageGeneration.test.ts
