+++ workers/.env.example
@@ -1,0 +1,31 @@
+# Cloudflare KV Namespace IDs
+CACHE_KV_ID=4b2c535505b7438ab4d382e170f3758e
+CACHE_KV_PREVIEW_ID=0214ee29e371415a9fa6ef0a9f229509
+
+SESSION_KV_ID=aa2e97aef4e14eb7a39872875b59ec74
+SESSION_KV_PREVIEW_ID=de725a4a38fc43f28cca9afd16c4d5f9
+
+RATE_LIMIT_KV_ID=0e0422b2b9f14a99b40adcb4906441e5
+RATE_LIMIT_KV_PREVIEW_ID=d0a6740231fd4448b5e84313dbea0129
+
+IMAGE_CACHE_KV_ID=83ea1a9180c84ed4a23b325583dbe737
+IMAGE_CACHE_KV_PREVIEW_ID=14ecb1acbc0348018c09a34efb215c7e
+
+# Production KV Namespace IDs
+PROD_CACHE_KV_ID=ac4cfe9a60374e0982097a6dda928044
+PROD_SESSION_KV_ID=bc2a774a46324cdab41089aa659ece23
+PROD_RATE_LIMIT_KV_ID=f4948d4c905944cbbd95aa798985179a
+PROD_IMAGE_CACHE_KV_ID=production_image_cache_namespace_id
+
+# R2 Configuration
+R2_BUCKET_NAME=prompt-builder
+R2_BUCKET_NAME_DEV=prompt-builder-dev
+R2_CUSTOM_DOMAIN=https://image.kantanprompt.com
+R2_S3_API_DEV=https://1b154d8dab68e47be1d8dc7734f1d802.r2.cloudflarestorage.com/prompt-builder-dev
+R2_S3_API_PROD=https://1b154d8dab68e47be1d8dc7734f1d802.r2.cloudflarestorage.com/prompt-builder
+
+# API Keys (Set these via wrangler secret put)
+# IMAGE_API_KEY=your-api-key
+# TRANSLATION_API_KEY=your-translation-api-key
+# R2_ACCESS_KEY_ID=your-r2-access-key
+# R2_SECRET_ACCESS_KEY=your-r2-secret-key
